# inventory/api-docs/openapi.yaml
openapi: 3.1.0
info:
  title: Green-Cart Inventory API
  version: 1.0.0
  description: |
    Inventory microservice for Green-Cart. Manages product inventory, stock levels,
    categories, and transaction history with support for low-stock alerts and bulk operations.
  contact:
    name: Green-Cart CTSE Team

servers:
  - url: http://localhost:8082
    description: Local development
  - url: https://inventory.greencart.example.com
    description: Production (Cloud Run)

tags:
  - name: Inventory
    description: Inventory item management
  - name: Categories
    description: Product category management
  - name: Health
    description: Service health

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      operationId: healthCheck
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /categories:
    post:
      tags: [Categories]
      summary: Create a new category
      operationId: createCategory
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryRequest"
      responses:
        "201":
          description: Category created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"

    get:
      tags: [Categories]
      summary: Get all categories
      operationId: getAllCategories
      parameters:
        - name: activeOnly
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Categories retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoriesListResponse"

  /categories/{id}:
    get:
      tags: [Categories]
      summary: Get category by ID
      operationId: getCategoryById
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "200":
          description: Category retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Categories]
      summary: Update category
      operationId: updateCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCategoryRequest"
      responses:
        "200":
          description: Category updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Categories]
      summary: Delete category
      operationId: deleteCategory
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CategoryId"
      responses:
        "204":
          description: Category deleted successfully
        "400":
          description: Category cannot be deleted (has associated items)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /categories/slug/{slug}:
    get:
      tags: [Categories]
      summary: Get category by slug
      operationId: getCategoryBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Category retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoryResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  /inventory:
    post:
      tags: [Inventory]
      summary: Create a new inventory item
      operationId: createItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateItemRequest"
      responses:
        "201":
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"

    get:
      tags: [Inventory]
      summary: Get all inventory items with filters
      operationId: getAllItems
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: sort
          in: query
          schema:
            type: string
            example: "-createdAt,name"
        - name: category
          in: query
          schema:
            type: string
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: inStock
          in: query
          schema:
            type: boolean
        - name: lowStock
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: sellerId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Items retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemsListResponse"

  /inventory/low-stock:
    get:
      tags: [Inventory]
      summary: Get low stock items
      operationId: getLowStockItems
      security:
        - bearerAuth: []
      parameters:
        - name: threshold
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Low stock items retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LowStockResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /inventory/{id}:
    get:
      tags: [Inventory]
      summary: Get inventory item by ID
      operationId: getItemById
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "200":
          description: Item retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Inventory]
      summary: Update inventory item
      operationId: updateItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateItemRequest"
      responses:
        "200":
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [Inventory]
      summary: Delete inventory item
      operationId: deleteItem
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ItemId"
      responses:
        "204":
          description: Item deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /inventory/sku/{sku}:
    get:
      tags: [Inventory]
      summary: Get inventory item by SKU
      operationId: getItemBySku
      parameters:
        - name: sku
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Item retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "404":
          $ref: "#/components/responses/NotFound"

  /inventory/{id}/stock:
    patch:
      tags: [Inventory]
      summary: Update item stock
      operationId: updateStock
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ItemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockUpdateRequest"
      responses:
        "200":
          description: Stock updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemResponse"
        "400":
          description: Insufficient stock or invalid adjustment
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  /inventory/{id}/transactions:
    get:
      tags: [Inventory]
      summary: Get transaction history for an item
      operationId: getTransactionHistory
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ItemId"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transaction history retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionHistoryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  /inventory/{id}/availability:
    get:
      tags: [Inventory]
      summary: Check item availability
      operationId: checkAvailability
      parameters:
        - $ref: "#/components/parameters/ItemId"
        - name: quantity
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Availability checked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
        "400":
          description: Invalid quantity
        "404":
          $ref: "#/components/responses/NotFound"

  /inventory/bulk-check-availability:
    post:
      tags: [Inventory]
      summary: Bulk check item availability
      operationId: bulkCheckAvailability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BulkAvailabilityRequest"
      responses:
        "200":
          description: Bulk availability checked successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkAvailabilityResponse"
        "400":
          description: Invalid request

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    ItemId:
      name: id
      in: path
      required: true
      schema:
        type: string
        example: "6613a1e2b9e4d50012abc123"

    CategoryId:
      name: id
      in: path
      required: true
      schema:
        type: string
        example: "6613a1e2b9e4d50012def456"

  schemas:
    HealthResponse:
      type: object
      required: [status, service]
      properties:
        status:
          type: string
          example: ok
        service:
          type: string
          example: inventory

    Category:
      type: object
      required: [_id, name, slug, isActive, createdAt, updatedAt]
      properties:
        _id:
          type: string
          example: "6613a1e2b9e4d50012def456"
        name:
          type: string
          example: "Vegetables"
        description:
          type: string
          example: "Fresh organic vegetables"
        icon:
          type: string
          example: "🥬"
        slug:
          type: string
          example: "vegetables"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    InventoryItem:
      type: object
      required: [_id, name, sku, category, price, stock, lowStockThreshold, unit, isActive, createdAt, updatedAt]
      properties:
        _id:
          type: string
          example: "6613a1e2b9e4d50012abc123"
        name:
          type: string
          example: "Organic Avocado (4 pack)"
        description:
          type: string
          example: "Hand-picked organic Hass avocados"
        sku:
          type: string
          example: "GC-AVD-001"
        category:
          $ref: "#/components/schemas/Category"
        price:
          type: number
          example: 890
        compareAtPrice:
          type: number
          example: 1100
        costPrice:
          type: number
          example: 600
        stock:
          type: integer
          example: 25
        lowStockThreshold:
          type: integer
          example: 10
        unit:
          type: string
          example: "Pack of 4"
        weight:
          type: number
          example: 500
        shelfLife:
          type: integer
          example: 7
        images:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
          example: ["Organic", "Non-GMO"]
        isActive:
          type: boolean
          example: true
        sellerId:
          type: string
          example: "merchant-123"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    Transaction:
      type: object
      required: [_id, item, type, quantity, previousStock, newStock, createdAt]
      properties:
        _id:
          type: string
        item:
          type: string
          example: "6613a1e2b9e4d50012abc123"
        type:
          type: string
          enum: [in, out, adjustment]
        quantity:
          type: integer
        previousStock:
          type: integer
        newStock:
          type: integer
        reason:
          type: string
        reference:
          type: string
        performedBy:
          type: string
        notes:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateCategoryRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
          example: "Vegetables"
        description:
          type: string
        icon:
          type: string
          example: "🥬"
        slug:
          type: string
          example: "vegetables"
        isActive:
          type: boolean
          default: true

    UpdateCategoryRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        slug:
          type: string
        isActive:
          type: boolean

    CreateItemRequest:
      type: object
      required: [name, sku, category, price, unit]
      properties:
        name:
          type: string
        description:
          type: string
        sku:
          type: string
        category:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
        costPrice:
          type: number
        stock:
          type: integer
          default: 0
        lowStockThreshold:
          type: integer
          default: 10
        unit:
          type: string
        weight:
          type: number
        shelfLife:
          type: integer
        images:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        isActive:
          type: boolean
          default: true

    UpdateItemRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
        costPrice:
          type: number
        lowStockThreshold:
          type: integer
        unit:
          type: string
        weight:
          type: number
        shelfLife:
          type: integer
        images:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        isActive:
          type: boolean

    StockUpdateRequest:
      type: object
      required: [quantity, type]
      properties:
        quantity:
          type: integer
        type:
          type: string
          enum: [in, out, adjustment]
        reason:
          type: string
        reference:
          type: string
        notes:
          type: string

    BulkAvailabilityRequest:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            type: object
            required: [id, quantity]
            properties:
              id:
                type: string
              quantity:
                type: integer

    CategoryResponse:
      type: object
      properties:
        category:
          $ref: "#/components/schemas/Category"

    CategoriesListResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"
        count:
          type: integer

    ItemResponse:
      type: object
      properties:
        item:
          $ref: "#/components/schemas/InventoryItem"

    ItemsListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer

    LowStockResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        count:
          type: integer

    TransactionHistoryResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer

    AvailabilityResponse:
      type: object
      properties:
        available:
          type: boolean
        currentStock:
          type: integer

    BulkAvailabilityResponse:
      type: object
      properties:
        results:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              available:
                type: boolean
              currentStock:
                type: integer
              error:
                type: string
                nullable: true

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string

  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    Forbidden:
      description: Admin access required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"